refactor(multipooler): relocate generic command and bootstrap-sentinel helpers out of backup files
Prep for extracting a backup/pgbackrest engine package: move helpers that aren't
backup-specific out of the rpc_backup*/rpc_first_backup files so those files
contain only backup logic.

- command_runner.go: runLongCommand (also used by rpc_check / expire / verify) and
  safeCombinedOutput — generic command execution, not backup-specific.
- bootstrap_sentinel.go: the bootstrapSentinel* family — a bootstrap/init concern
  (used by the first-backup orchestration and manager.go), not the backup engine.

Pure within-package relocation; no behavior change.
